Task3/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
torch>=2.2.0
transformers>=4.33.0
huggingface-hub>=0.19.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
